'Nimrod' is now called 'Nim'
I fixed Nim's URL and rename Nimrod to Nim in README.
I think the name of this plugin should be 'nim.vim'.
But as you know PR can't change the repository name.
I remained this plugin's name, 'nimrod.vim' as it is
not to confuse users.
When the name of this plugin is changed, they should
be replaced.